[MOV] f&r: move find & replace plugin to store
Task: 3994472
