+++ src/components/side_panel/find_and_replace/find_and_replace_store.ts
@@ -2,6 +2,7 @@ import { debounce, getSearchRegex, isInside, positionToZone } from "../../../hel
 import { HighlightProvider, HighlightStore } from "../../../stores/highlight_store";
 import { CellPosition, Color, Command, Highlight } from "../../../types";
 
+import { canonicalizeNumberContent } from "../../../helpers/locale";
 import { Get } from "../../../store_engine";
 import { SpreadsheetStore } from "../../../stores";
 import { SearchOptions } from "../../../types/find_and_replace";
@@ -113,6 +114,10 @@ export class FindAndReplaceStore extends SpreadsheetStore implements HighlightPr
       case "ACTIVATE_SHEET":
         this.isSearchDirty = true;
         break;
+      case "REPLACE_SEARCH":
+        for (const match of cmd.matches) {
+          this.replaceMatch(match, cmd.searchString, cmd.replaceWith, cmd.searchOptions);
+        }
     }
   }
 
@@ -305,6 +310,32 @@ export class FindAndReplaceStore extends SpreadsheetStore implements HighlightPr
       matches: this.searchMatches,
       searchOptions: this.searchOptions,
     });
+  }
+
+  private replaceMatch(
+    selectedMatch: CellPosition,
+    searchString: string,
+    replaceWith: string,
+    searchOptions: SearchOptions
+  ) {
+    const cell = this.getters.getCell(selectedMatch);
+    if (!cell?.content) {
+      return;
+    }
+
+    if (cell?.isFormula && !searchOptions.searchFormulas) {
+      return;
+    }
+
+    const searchRegex = getSearchRegex(searchString, searchOptions);
+    const replaceRegex = new RegExp(searchRegex.source, searchRegex.flags + "g");
+    const toReplace: string | null = this.getters.getCellText(
+      selectedMatch,
+      searchOptions.searchFormulas
+    );
+    const content = toReplace.replace(replaceRegex, replaceWith);
+    const canonicalContent = canonicalizeNumberContent(content, this.getters.getLocale());
+    this.model.dispatch("UPDATE_CELL", { ...selectedMatch, content: canonicalContent });
   }
 
   private getSearchableString(position: CellPosition): string {
