+++ src/plugins/ui_feature/index.ts
@@ -2,7 +2,6 @@ export * from "./autofill";
 export * from "./automatic_sum";
 export * from "./collaborative";
 export * from "./data_cleanup";
-export * from "./find_and_replace";
 export * from "./format";
 export * from "./header_visibility_ui";
 export * from "./insert_pivot";
